requirements.txt
faiss-cpu
fastapi
uvicorn
python-docx
requests
